Bug fix: Restart flutter-pi after slide deletion
If there is a file to be deleted on pi but not nothing to be downloaded,
rclone will print "There was nothing to transfer" and our script will
not kill flutter-pi. As a result, a deleted file on WebDAV remains
displayed on the digital signage.

This commit fix the bug by always "diff" two directories.